# Fix some CSS case sensitivity issues