- Case should not matter when parsing CSS declaration properties ("`Font-Weight`"), named values ("`Bold`"), named length ("`LARGER`"), and length units ("`10PX`").
- Case should matter when matching class names ("`<div class="bottomBorder"` should not use CSS defined as "div.bottomborder").
- Fix color rendering of background image (the conversion of internal images from 32bpp to the final 16bpp was simply not implemented).
- Fix existing unwanted lowercasing of background image file reference (`url("BackgrounImage.jpeg")` was lowercased in the process of parsing it.

See discussion in #140. Closes #140.

There is still an existing issue with resolution of relative image path there that could fail (couldn't find a quick fix for this), and some other existing strange issue when no-repeat is used (as in the After screenshot before, that looked the same before, with a lowercase filename and fullpath).

Sample file for testing that kind of stuff: 

[test-css-nocase.epub.zip](https://github.com/koreader/crengine/files/1946521/test-css-nocase.epub.zip)

Before:
<kbd>![before](https://user-images.githubusercontent.com/24273478/39239730-171e9e18-4882-11e8-83a5-935bc50135c7.png)</kbd>

After:
<kbd>![image](https://user-images.githubusercontent.com/24273478/39239755-2aece5b2-4882-11e8-906a-2663a9988678.png)</kbd>