@poire-z 2018-04-25
Btw, I wonder about something while testing.
say we have `background-image: url("../paper.png")` in `/OEBPS/template.css`
say we have `/OEBPS/Text/Chapter1.html` that includes `<style link="../templace.css">`
Where should the relative `../paper.png` resolve from ?
- the templace.css location, so that would give `/paper.png` ?
- or the Chapter1.html location, so that would give `/OEBPS/paper.png`