@Frenzie 2018-04-25
The CSS file shouldn't care about where it's called from. If it did it'd be unusable. That's `/paper.png`.