@poire-z 2018-04-25
Some timings, with the same book I timed in https://github.com/koreader/crengine/pull/97#issuecomment-369733864.

Before #163:

cache before | load time | render time
---------------- | -------------------------- | ----------------
no cache yet | 24 s | 23s

After #163:

cache before | load time | render time
---------------- | -------------------------- | ----------------
no cache yet | 18 s | 16s

So, getting rid of that `lowercase()` in classname checks, as predicted in https://github.com/koreader/koreader/issues/3459#issuecomment-375909783, gives a good 25%/30% gain !