@poire-z 2018-04-26
On the emulator.
(Same book takes ~145s on my Kobo, so I expect 110s, i'll see when it's bump in a nightly)