@poire-z 2018-04-28
145s > 99s on my kobo.
And no problem on arm kobo or android with that endianess worry.